refactor(number): separate `integer` as a module
