+++ drake-lexer/src/number/tests.rs
@@ -1,22 +1,7 @@
-use core::fmt::Debug;
-use core::str::Chars;
-
 use futures_executor::block_on;
 use somen::prelude::*;
-use somen::stream::rewind::BufferedRewinder;
-use somen::stream::{InfallibleStream, IteratorStream};
 
-#[test]
-fn integer() {
-    block_on(async {
-        let parser = &mut super::integer().complete();
-        assert_parser(parser, "42", 42).await;
-        assert_parser(parser, "0xDEADBEEF", 0xDEADBEEF).await;
-        assert_parser(parser, "0xcafebabe", 0xcafebabe).await;
-        assert_parser(parser, "0o644", 0o644).await;
-        assert_parser(parser, "0b01010110", 0b01010110).await;
-    })
-}
+use crate::utils::{assert_parser, assert_parser_fail};
 
 #[test]
 fn digits() {
@@ -50,22 +35,3 @@ fn digits_trailing_zeros() {
     });
 }
 
-async fn assert_parser<T: PartialEq + Debug>(
-    parser: &mut impl Parser<
-        BufferedRewinder<InfallibleStream<IteratorStream<Chars<'static>>>>,
-        Output = T,
-    >,
-    s: &'static str,
-    res: T,
-) {
-    let mut stream = stream::from_iter(s.chars()).buffered_rewind();
-    assert_eq!(parser.parse(&mut stream).await, Ok(res));
-}
-
-async fn assert_parser_fail(
-    parser: &mut impl Parser<BufferedRewinder<InfallibleStream<IteratorStream<Chars<'static>>>>>,
-    s: &'static str,
-) {
-    let mut stream = stream::from_iter(s.chars()).buffered_rewind();
-    assert!(parser.parse(&mut stream).await.is_err());
-}
